combine tester env/flag files
